% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adr}
\alias{adr}
\title{Administrative data}
\format{
\subsection{\code{adr}}{

A data frame with 1000 rows and 5 columns:
\describe{
\item{gcse_score}{Child's GCSE score (at approximately age 16 years), as a
percentage, calculated from the total of an individual’s top eight GCSE
qualifications, ranked in terms of points}
\item{log_income}{Logarithm of the family income per annum, in UK pounds,
when the child is aged 5 years}
\item{mated}{Mother's educational level: post-16 years qualification or
not}
\item{ks2_score}{Child's Key Stage 2 score (at approximately age 11 years),
as a percentage, calculated from the total of an individual's English,
maths and science scores}
\item{r_cra}{Missingness indicator: whether a complete case or not}
...
}
}
}
\usage{
adr
}
\description{
A simulated dataset
}
\keyword{datasets}
