% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmi}
\alias{bmi}
\title{Child body mass index data}
\format{
\subsection{\code{bmi} A data frame with 1000 rows and 6 columns:}{

\describe{
\item{bmi7}{Child's body mass index at age 7 years}
\item{matage}{Mother's age at first pregnancy, standardised relative to a
mean age of 30 years}
\item{mated}{Mother's educational level: post-16 years qualification or
not}
\item{pregsize}{Mother's pregnancy size: singleton or twins}
\item{bwt}{Child's birth weight in kilograms}
\item{r}{Missingness indicator: whether bmi7 is reported or not}
...
}
}
}
\usage{
bmi
}
\description{
A simulated dataset
}
\keyword{datasets}
