% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descMissData.R
\name{descMissData}
\alias{descMissData}
\title{Lists missing data patterns in the specified dataset}
\usage{
descMissData(y, covs, data)
}
\arguments{
\item{y}{The analysis model outcome, specified as a string}

\item{covs}{The analysis model covariate(s), specified as a string (space
delimited)}

\item{data}{A data frame containing the specified analysis model outcome and
covariate(s)}
}
\value{
A summary of the missing data patterns
}
\description{
This function summarises the missing data patterns in the specified dataset.
Each row in the output corresponds to a missing data pattern (1=observed,
0=missing). The number and percentage of observations is also displayed for
each missing data pattern.
}
\examples{
descMissData(y="bmi7", covs="matage mated", data=bmi)
descMissData(y="bmi7", covs="matage mated pregsize bwt", data=bmi)
}
