% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp}
\alias{fp}
\title{Fractional Polynomial Transformation as per {mfp} package: Ambler G, Benner A (2022). mfp: Multivariable Fractional Polynomials. R package version 1.5.2.2, \url{https://CRAN.R-project.org/package=mfp}.}
\usage{
fp(x, df = 4, select = NA, alpha = NA, scale = TRUE)
}
\arguments{
\item{x}{input variable}

\item{df}{degrees of freedom. df = 4 corresponds to an FP model with maximum of 2 permitted degrees of freedom}

\item{select}{sets the variable selection level}

\item{alpha}{sets the FP selection level}

\item{scale}{use pre-transformation scaling}
}
\description{
Fractional Polynomial Transformation as per {mfp} package: Ambler G, Benner A (2022). mfp: Multivariable Fractional Polynomials. R package version 1.5.2.2, \url{https://CRAN.R-project.org/package=mfp}.
}
