% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midocdemo.R
\name{midocdemo}
\alias{midocdemo}
\title{Run an interactive vignette for the midoc package}
\usage{
midocdemo()
}
\description{
Runs an interactive version of the midoc vignette: Multiple Imputation DOCtor
(midoc). In the interactive version, you can apply midoc functions using your
own DAG and data.
}
