% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midocshinyapp.R
\name{midocShinyApp}
\alias{midocShinyApp}
\title{Run a browser-based version of the midoc package}
\usage{
midocShinyApp()
}
\value{
A browser-based version of the midoc package
}
\description{
Runs a browser-based version of the midoc package. In this version, you can
explore midoc functions using exemplar datasets or apply midoc functions
using your own DAG and data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Run the browser-based version of the midoc package
midocShinyApp()
\dontshow{\}) # examplesIf}
}
