% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qol}
\alias{qol}
\title{Randomised controlled trial data}
\format{
\subsection{\code{qol}}{

A data frame with 1000 rows and 6 columns:
\describe{
\item{group}{Randomisation group: 1 = Placebo, 2 = Active treatment}
\item{age0}{Participant's age at randomisation (baseline), in years}
\item{qol0}{Participant's quality of life at randomisation (baseline),
measured using the EuroQol Visual Analogue Scale (EQ-VAS)}
\item{qol3}{Participant's quality of life at 3 months post-randomisation,
measured using EQ-VAS}
\item{qol12}{Participant's quality of life at 12 months post-randomisation,
measured using EQ-VAS}
\item{r_qol12}{Missingness indicator: whether qol12 is reported or not}
}
}
}
\usage{
qol
}
\description{
A simulated dataset
}
\keyword{datasets}
