% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summMissData.R
\name{summMissData}
\alias{summMissData}
\title{Summarise the data distribution for complete and incomplete records}
\usage{
summMissData(
  y,
  r_cra,
  data,
  covs = NULL,
  by = NULL,
  plot = TRUE,
  plotprompt = TRUE,
  message = TRUE,
  ...
)
}
\arguments{
\item{y}{The analysis model outcome variable(s), specified as a string (space
delimited) or a list}

\item{r_cra}{The complete record indicator variable, which must be a factor}

\item{data}{A data frame containing the specified analysis model outcome,
covariate(s), and if specified, stratification variable(s)}

\item{covs}{Optional analysis model covariate(s), specified as a string (space
delimited) or a list}

\item{by}{Optional additional stratification variable(s); if specified, the
data are subsetted according to both the complete record indicator and the
values of the stratification variable(s) and summary data are displayed for
each subset in turn}

\item{plot}{If TRUE (the default), summary plots are displayed; note that
stratification variables are ignored in the plot; use plot = FALSE to
disable the plots}

\item{plotprompt}{If TRUE (the default), the user is prompted before each plot
is displayed; use plotprompt=FALSE to remove the prompt}

\item{message}{If TRUE (the default), displays a table of descriptive
statistics comparing complete and incomplete records; use message = FALSE to
suppress the message}

\item{...}{Further arguments passed to \link[table1]{table1}}
}
\value{
A summary of descriptive statistics comparing complete and incomplete
records and, optionally, descriptive plots.
}
\description{
Produces a table of descriptive statistics for the analysis model outcome
('y') and covariate ('covs') variables, stratified by the complete record
indicator ('r_cra') and optionally, by other stratification variable(s).
Alternatively, 'y' can indicate the primary variable of interest, e.g. 'y'
could refer to an exposure or intervention, with all other variables listed in
'covs'. The table displays the percentage of missing values for each variable,
plus the mean and standard deviation of numeric variables, and frequency and
percentage of categorical variables. Optional plots display box plots for
numeric variables and stacked bar charts for categorical variables.
}
\details{
The summary table is created using the \link[table1]{table1} function. Further
arguments can be passed to this function - see the help file for
\link[table1]{table1} for more information. Note that the 'transpose' feature
is not available. The summary table and plots can be used to compare the
characteristics of complete and incomplete records. Note that there may be
differences in variable distributions between complete and incomplete records
if data are missing at random or missing not at random. Therefore, any
observed differences cannot be used as evidence for or against the data
missing at random assumption.
}
\examples{
summMissData(y="bmi7", r_cra="r", data=bmi, covs="matage mated bwt pregsize")
}
