% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcra.R
\name{checkcra}
\alias{checkcra}
\title{Check complete records analysis is valid}
\usage{
checkcra(y, covs, r, mdag)
}
\arguments{
\item{y}{the analysis model outcome, specified as a string}

\item{covs}{the analysis model covariate(s), specified as a string (space delimited)}

\item{r}{the complete record indicator, specified as a string}

\item{mdag}{the directed acyclic graph specified as a string using "dagitty" syntax, including all observed and
unobserved variables related to the analysis model variables and their missingness}
}
\value{
A message indicating whether complete records analysis is valid under the proposed missingness
DAG and analysis model outcome and covariate(s)
}
\description{
Check complete records analysis is valid
}
\examples{
checkcra(y="edscore14", covs="iq8", r="r", mdag="iq8 -> edscore14 iq8 -> r")
checkcra(y="edscore14", covs="iq8", r="r",
 mdag="smoking -> iq8 -> edscore14 smoking -> edscore14 smoking -> r")
checkcra(y="edscore14", covs="iq8 smoking", r="r",
 mdag="smoking -> iq8 -> edscore14 smoking -> edscore14 smoking -> r")
}
