% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcra.R
\name{checkCRA}
\alias{checkCRA}
\title{Check complete records analysis is valid}
\usage{
checkCRA(y, covs, r, mdag)
}
\arguments{
\item{y}{the analysis model outcome, specified as a string}

\item{covs}{the analysis model covariate(s), specified as a string (space delimited)}

\item{r}{the complete record indicator, specified as a string}

\item{mdag}{the directed acyclic graph specified as a string using "dagitty" syntax, including all observed and
unobserved variables related to the analysis model variables and their missingness}
}
\value{
A message indicating whether complete records analysis is valid under the proposed missingness
DAG and analysis model outcome and covariate(s)
}
\description{
Check complete records analysis is valid
}
\examples{
checkCRA(y="bmi7", covs="matage", r="R",
 mdag="mated -> bmi7 mated -> matage matage -> bmi7 mated -> R")
checkCRA(y="bmi7", covs="matage mated", r="R",
 mdag="mated -> bmi7 mated -> matage matage -> bmi7 mated -> R")
checkCRA(y="bmi7", covs="matage mated", r="R",
 mdag="mated -> bmi7 mated -> matage matage -> bmi7 mated -> R bmi7 -> R")
}
