% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmi.R
\name{checkmi}
\alias{checkmi}
\title{Check multiple imputation is valid}
\usage{
checkmi(dep, preds, r, mdag)
}
\arguments{
\item{dep}{the partially observed variable, specified as a string}

\item{preds}{the imputation model predictor(s), specified as a string (space delimited)}

\item{r}{the partially observed variable's missingness indicator, specified as a string}

\item{mdag}{the directed acyclic graph specified as a string using "dagitty" syntax, including all observed and
unobserved variables related to the analysis model variables and their missingness}
}
\value{
A message indicating whether multiple imputation is valid under the proposed missingness
DAG and imputation model
}
\description{
Check multiple imputation is valid
}
\examples{
checkmi(dep="iq8", preds="edscore14 smoking", r="r",
 mdag="smoking -> iq8 -> edscore14 smoking -> edscore14 smoking -> r")
checkmi(dep="iq8", preds="edscore14", r="r",
 mdag="smoking -> iq8 -> edscore14 smoking -> edscore14 smoking -> r")
}
