% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmi.R
\name{checkMI}
\alias{checkMI}
\title{Check multiple imputation is valid}
\usage{
checkMI(dep, preds, r_dep, mdag)
}
\arguments{
\item{dep}{the partially observed variable, specified as a string}

\item{preds}{the imputation model predictor(s), specified as a string (space delimited)}

\item{r_dep}{the partially observed variable's missingness indicator, specified as a string}

\item{mdag}{the directed acyclic graph specified as a string using "dagitty" syntax, including all observed and
unobserved variables related to the analysis model variables and their missingness}
}
\value{
A message indicating whether multiple imputation is valid under the proposed missingness
DAG and imputation model
}
\description{
Check multiple imputation is valid
}
\examples{
checkMI(dep="bmi7", preds="matage mated pregsize", r_dep="R",
 mdag="mated -> bmi7 mated -> matage matage -> bmi7 mated -> R mated -> bwt
 pregsize -> bmi7 pregsize -> bwt  U -> bwt  U -> R  U -> mated")
checkMI(dep="bmi7", preds="matage mated bwt", r_dep="R",
 mdag="mated -> bmi7 mated -> matage matage -> bmi7 mated -> R mated -> bwt
 pregsize -> bmi7 pregsize -> bwt  U -> bwt  U -> R  U -> mated")
checkMI(dep="matage", preds="bmi7 mated bwt", r_dep="R",
 mdag="mated -> bmi7 mated -> matage matage -> bmi7 mated -> R mated -> bwt
 pregsize -> bmi7 pregsize -> bwt  U -> bwt  U -> R  U -> mated")
}
