% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmodspec.R
\name{checkModSpec}
\alias{checkModSpec}
\title{Specify and inspect parametric model specification}
\usage{
checkModSpec(
  formula,
  family,
  by = NULL,
  data = NULL,
  plot = TRUE,
  message = TRUE
)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted, with the
dependent variable on the left of a ~ operator, and the covariates,
separated by + operators, on the right, specified as a string}

\item{family}{A description of the error distribution and link function to be
used in the model, specified as a string; family functions that are
supported are "gaussian(identity)" and "binomial(logit)"}

\item{by}{Optional stratification variable(s), specified as a string (space
delimited) or a list of factors; if specified, the parametric model will be
fit for each subset of the data determined by the values of the factor(s)}

\item{data}{Optionally, a data frame containing all the variables stated in
the formula and if specified, stratification variable(s)}

\item{plot}{If TRUE (the default), and a dataset is supplied, displays a plot
which can be used to explore the form of the specified model; note that
stratification variables are ignored in the plot; use plot = FALSE to
disable the plot}

\item{message}{If TRUE (the default), and a dataset is supplied, displays a
message indicating whether the relationships between the dependent variable
and covariates are likely to be correctly specified or not; use message =
FALSE to suppress the message}
}
\value{
An object of type 'mimod' (a list containing the specified formula,
family, and, if specified, dataset name). Optionally, if required and a
dataset is supplied, a message indicating whether the relationships between
the dependent variable and covariates are likely to be correctly specified
or not.
}
\description{
Specify a parametric model (which may represent the analysis or imputation
model). Optionally, if a dataset is supplied, explore whether the observed
relationships in the specified dataset are consistent with the proposed
parametric model.
}
\examples{
# Example (incorrectly) assuming a linear relationship
checkModSpec(formula="bmi7~matage+mated+pregsize",
             family="gaussian(identity)", data=bmi)
  ## For the example above, (correctly) assuming a quadratic relationship
checkModSpec(formula="bmi7~matage+I(matage^2)+mated+pregsize",
             family="gaussian(identity)", data=bmi)
}
\references{
Curnow E, Carpenter JR, Heron JE, et al. 2023. Multiple
imputation of missing data under missing at random: compatible imputation
models are not sufficient to avoid bias if they are mis-specified. J Clin
Epidemiol. \url{doi:10.1016/j.jclinepi.2023.06.011}
}
