% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmodspec.R
\name{checkmodspec}
\alias{checkmodspec}
\title{Check specification of the proposed model (analysis or imputation) using the observed data}
\usage{
checkmodspec(formula, family, data)
}
\arguments{
\item{formula}{a formula object: a symbolic description of the model to be fitted, with the model outcome on the left of a ~ operator, and the covariates, separated by + operators, on the right.}

\item{family}{a description of the error distribution and link function to be used in the model. Family functions that are supported are gaussian(identity) and binomial(logit).}

\item{data}{a data frame containing all the variables stated in the formula.}
}
\value{
A message indicating whether the relationships between the model outcome and covariates are likely to be correctly specified or not, plus results of a method for examining model mis-specification.
}
\description{
Check specification of the proposed model (analysis or imputation) using the observed data
}
\examples{
checkmodspec(formula=bmi7~matage+mated+pregsize, family=gaussian(identity), data=bmi)
checkmodspec(formula=bmi7~matage+I(matage^2)+mated+pregsize, family=gaussian(identity), data=bmi)
checkmodspec(formula=mated~bmi7+matage, family=binomial(logit), data=bmi)
checkmodspec(formula=mated~bmi7+matage+I(matage^2), family=binomial(logit), data=bmi)
}
