% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmodspec.R
\name{checkModSpec}
\alias{checkModSpec}
\title{Inspect parametric model specification}
\usage{
checkModSpec(formula, family, data, modplot = TRUE)
}
\arguments{
\item{formula}{A formula object: a symbolic description of the model to be
fitted, with the dependent variable on the left of a ~ operator, and the
covariates, separated by + operators, on the right}

\item{family}{A description of the error distribution and link function to be
used in the model; family functions that are supported are
gaussian(identity) and binomial(logit)}

\item{data}{A data frame containing all the variables stated in the formula}

\item{modplot}{If TRUE and there is evidence of model mis-specification,
displays a plot which can be used to explore the functional form of each
covariate in the specified model; use modplot = FALSE to disable the plot}
}
\value{
A message indicating whether the relationships between the dependent
variable and covariates are likely to be correctly specified or not, plus
an object of type 'miprop' (a list containing the specified formula,
family, and dataset name)
}
\description{
Explore whether the observed relationships in the specified dataset are
consistent with the proposed parametric model (which may represent the
analysis or imputation model).
}
\examples{
checkModSpec(formula=bmi7~matage+mated, family=gaussian(identity), data=bmi)
checkModSpec(formula=bmi7~matage+I(matage^2)+mated+pregsize,
  family=gaussian(identity), data=bmi)

checkModSpec(formula=mated~bmi7+matage, family=binomial(logit), data=bmi)
checkModSpec(formula=mated~bmi7+matage+I(matage^2),
  family=binomial(logit), data=bmi)
}
