% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doMImice.R
\name{doMImice}
\alias{doMImice}
\title{Performs multiple imputation}
\usage{
doMImice(mipropobj, seed, substmod = " ")
}
\arguments{
\item{mipropobj}{An object of type 'miprop', created by a call to 'proposeMI'}

\item{seed}{An integer that is used to set the seed of the 'mice' call}

\item{substmod}{Optionally, a symbolic description of the substantive model
to be fitted, specified as a string; if supplied, the model will be fitted
to each imputed dataset and the results pooled}
}
\value{
A 'mice' object of type 'mids'
}
\description{
Performs multiple imputation using \link[mice]{mice}, based on the 'mice'
options and dataset specified by a call to  \link[midoc]{proposeMI}.
}
\examples{
mimod <- checkModSpec(formula="bmi7~matage+I(matage^2)+mated+pregsize",
  family="gaussian(identity)", data=bmi)
miprop <- proposeMI(mimod, data=bmi)

imp <- doMImice(miprop,123)

doMImice(miprop, 123, substmod="lm(bmi7 ~ matage + I(matage^2) + mated)")
}
