% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midoc-package.R
\docType{package}
\name{midoc-package}
\alias{midoc}
\alias{midoc-package}
\title{midoc: A Decision-Making System For Multiple Imputation}
\description{
A guidance system for analysis with missing data. It incorporates expert, up-to-date methodology to help researchers choose the most appropriate analysis approach when some data are missing. You provide the available data and the assumed causal structure, including the likely causes of missing data. 'midoc' will advise whether multiple imputation is needed, and if so, how best to perform it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://elliecurnow.github.io/midoc/}
}

}
\author{
\strong{Maintainer}: Elinor Curnow \email{elinor.curnow@bristol.ac.uk} (\href{https://orcid.org/0000-0002-3109-3647}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jon Heron
  \item Rosie Cornish
  \item Kate Tilling
  \item James Carpenter
}

}
\keyword{internal}
