% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposeMI.R
\name{proposeMI}
\alias{proposeMI}
\title{Suggests multiple imputation options}
\usage{
proposeMI(
  mimodobj,
  prop_complete = NA,
  data = NULL,
  plot = TRUE,
  plotprompt = TRUE,
  message = TRUE
)
}
\arguments{
\item{mimodobj}{An object, or list of objects, of type 'mimod', which stands
for 'multiple imputation model', created by a call to
\link[midoc]{checkModSpec}}

\item{prop_complete}{Optionally, the proportion of complete records,
specified as a number between 0 and 1. This is only required if a dataset
is not specified. If a dataset is specified, the proportion of complete
records will be calculated based on all columns included in the dataset.}

\item{data}{Optionally, a data frame containing all the variables required
for imputation and the substantive analysis; if stratification variable(s)
are included in the 'mimod' object(s), these will be carried over to 'midoc'
functions 'doMImice' and 'doMNARmice' and multiple imputation will be
performed for each subset of the data determined by the values of the
stratification variable(s)}

\item{plot}{If TRUE (the default), and a dataset is supplied, displays
diagnostic plots for the proposed 'mice' call; use plot=FALSE to disable
the plots}

\item{plotprompt}{If TRUE (the default), and a dataset is supplied, the user
is prompted before the second plot is displayed; use plotprompt=FALSE to
remove the prompt and display all plots at the same time}

\item{message}{If TRUE (the default), displays a message describing the
proposed 'mice' options; use message=FALSE to suppress the message}
}
\value{
An object of type 'miprop', which can be used to run 'mice' using
the proposed options, plus optionally, a message and, if a dataset is
supplied, diagnostic plots describing the proposed 'mice' options
}
\description{
Suggests the \link[mice]{mice} options to perform multiple imputation, based
on the set of imputation models (one for each partially observed variable)
specified by calls to \link[midoc]{checkModSpec} and the proportion of
complete records. Optionally, if a dataset is supplied, diagnostic plots are
created based on the proposed 'mice' options.
}
\examples{
# First specify each imputation model as a 'mimod' object
## (suppressing the message)
mimod_bmi7 <- checkModSpec(formula="bmi7~matage+I(matage^2)+mated+pregsize",
                           family="gaussian(identity)",
                           data=bmi,
                           message=FALSE)
mimod_pregsize <- checkModSpec(
                           formula="pregsize~bmi7+matage+I(matage^2)+mated",
                           family="binomial(logit)",
                           data=bmi,
                           message=FALSE)

# Display the proposed 'mice' options
## When specifying a single imputation model
proposeMI(mimodobj=mimod_bmi7,
                   data=bmi)
## When specifying more than one imputation model (suppressing the plots)
proposeMI(mimodobj=list(mimod_bmi7,mimod_pregsize),
                    data=bmi,
                    plot=FALSE)
}
