% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposeMI.R
\name{proposeMI}
\alias{proposeMI}
\title{Suggests multiple imputation options}
\usage{
proposeMI(mimodobj, data, message = TRUE, diagplot = TRUE, printprompt = TRUE)
}
\arguments{
\item{mimodobj}{An object of type 'mimod', which stands for 'multiple
imputation model', created by a call to \link[midoc]{checkModSpec}}

\item{data}{A data frame containing all the variables required for the
imputation and substantive analysis models}

\item{message}{If TRUE (the default), displays a message describing the
proposed 'mice' options; use message=FALSE to suppress the message}

\item{diagplot}{If TRUE (the default), displays diagnostic plots for the
proposed 'mice' call; use diagplot=FALSE to disable the plots}

\item{printprompt}{If TRUE (the default), the user is prompted before the
second plot is displayed; use printprompt=FALSE to remove the prompt}
}
\value{
An object of type 'miprop', which can be used to run 'mice' using the
proposed options, plus, optionally, a message and diagnostic plots describing
the proposed 'mice' options
}
\description{
Suggests the optimal \link[mice]{mice} options to perform multiple
imputation, based on the proposed imputation model and specified dataset.
}
\examples{
mimod <- checkModSpec(formula="bmi7~matage+I(matage^2)+mated+pregsize",
family="gaussian(identity)", data=bmi)
proposeMI(mimod, data=bmi)
}
